requirements.txt
Django==4.0.3
Requests==2.28.1
Python-Dateutil==2.8.2